# openapi.yaml
# Сгенерировано из FastAPI. Обновлять при изменении API (Definition of Done, CLAUDE.md).
openapi: 3.1.0
info:
  title: Совфрахт Детали — API
  description: 'Распознавание судовых деталей по фото: закупка и оценка ремонта.'
  version: 0.1.0
paths:
  /api/v1/healthz:
    get:
      tags:
      - service
      summary: Живость сервиса и доступность БД
      operationId: healthz_api_v1_healthz_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Healthz Api V1 Healthz Get
  /api/v1/auth/token:
    post:
      tags:
      - auth
      summary: Вход по логину и паролю
      operationId: login_api_v1_auth_token_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/refresh:
    post:
      tags:
      - auth
      summary: Обновление access-токена
      operationId: refresh_api_v1_auth_refresh_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessTokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/me:
    get:
      tags:
      - auth
      summary: Текущий пользователь и роль
      operationId: me_api_v1_auth_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
  /api/v1/users:
    get:
      tags:
      - users
      summary: Список пользователей организации
      operationId: list_users_api_v1_users_get
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPage'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - users
      summary: Создать пользователя
      operationId: create_user_api_v1_users_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/users/{user_id}:
    patch:
      tags:
      - users
      summary: Изменить пользователя
      operationId: update_user_api_v1_users__user_id__patch
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/vessels:
    get:
      tags:
      - vessels
      summary: Суда организации
      operationId: list_vessels_api_v1_vessels_get
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VesselPage'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - vessels
      summary: Добавить судно
      operationId: create_vessel_api_v1_vessels_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VesselCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VesselRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/vessels/{vessel_id}:
    patch:
      tags:
      - vessels
      summary: Изменить судно
      operationId: update_vessel_api_v1_vessels__vessel_id__patch
      parameters:
      - name: vessel_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Vessel Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VesselUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VesselRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    AccessTokenResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        token_type:
          type: string
          title: Token Type
          default: bearer
        expires_in:
          type: integer
          title: Expires In
      type: object
      required:
      - access_token
      - expires_in
      title: AccessTokenResponse
      description: Ответ POST /auth/refresh — только новый access.
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    LoginRequest:
      properties:
        login:
          type: string
          maxLength: 255
          minLength: 1
          title: Login
        password:
          type: string
          maxLength: 256
          minLength: 1
          title: Password
      type: object
      required:
      - login
      - password
      title: LoginRequest
    RefreshRequest:
      properties:
        refresh_token:
          type: string
          title: Refresh Token
      type: object
      required:
      - refresh_token
      title: RefreshRequest
    Role:
      type: string
      enum:
      - mechanic
      - supplier_manager
      - fleet_owner
      - expert
      - admin
      title: Role
      description: Роли RBAC (FR-AUTH-02, NFR-SEC-03).
    TokenResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        refresh_token:
          type: string
          title: Refresh Token
        token_type:
          type: string
          title: Token Type
          default: bearer
        expires_in:
          type: integer
          title: Expires In
        user:
          $ref: '#/components/schemas/UserRead'
      type: object
      required:
      - access_token
      - refresh_token
      - expires_in
      - user
      title: TokenResponse
      description: Ответ POST /auth/token (docs/08 §2).
    UserCreate:
      properties:
        login:
          type: string
          maxLength: 255
          minLength: 3
          title: Login
        full_name:
          type: string
          maxLength: 255
          minLength: 1
          title: Full Name
        password:
          type: string
          maxLength: 256
          minLength: 8
          title: Password
        role:
          $ref: '#/components/schemas/Role'
        vessel_ids:
          items:
            type: string
            format: uuid
          type: array
          title: Vessel Ids
      type: object
      required:
      - login
      - full_name
      - password
      - role
      title: UserCreate
    UserPage:
      properties:
        items:
          items:
            $ref: '#/components/schemas/UserRead'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        page_size:
          type: integer
          title: Page Size
      type: object
      required:
      - items
      - total
      - page
      - page_size
      title: UserPage
    UserRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        organization_id:
          type: string
          format: uuid
          title: Organization Id
        login:
          type: string
          title: Login
        full_name:
          type: string
          title: Full Name
        role:
          $ref: '#/components/schemas/Role'
        is_active:
          type: boolean
          title: Is Active
        vessels:
          items:
            $ref: '#/components/schemas/VesselBrief'
          type: array
          title: Vessels
      type: object
      required:
      - id
      - organization_id
      - login
      - full_name
      - role
      - is_active
      title: UserRead
    UserUpdate:
      properties:
        full_name:
          anyOf:
          - type: string
            maxLength: 255
            minLength: 1
          - type: 'null'
          title: Full Name
        password:
          anyOf:
          - type: string
            maxLength: 256
            minLength: 8
          - type: 'null'
          title: Password
        role:
          anyOf:
          - $ref: '#/components/schemas/Role'
          - type: 'null'
        is_active:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Active
        vessel_ids:
          anyOf:
          - items:
              type: string
              format: uuid
            type: array
          - type: 'null'
          title: Vessel Ids
      type: object
      title: UserUpdate
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    VesselBrief:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        imo:
          anyOf:
          - type: string
          - type: 'null'
          title: Imo
      type: object
      required:
      - id
      - name
      title: VesselBrief
    VesselCreate:
      properties:
        name:
          type: string
          maxLength: 255
          minLength: 1
          title: Name
        imo:
          anyOf:
          - type: string
            maxLength: 32
          - type: 'null'
          title: Imo
        type:
          anyOf:
          - type: string
            maxLength: 128
          - type: 'null'
          title: Type
      type: object
      required:
      - name
      title: VesselCreate
    VesselPage:
      properties:
        items:
          items:
            $ref: '#/components/schemas/VesselRead'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        page_size:
          type: integer
          title: Page Size
      type: object
      required:
      - items
      - total
      - page
      - page_size
      title: VesselPage
    VesselRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        organization_id:
          type: string
          format: uuid
          title: Organization Id
        name:
          type: string
          title: Name
        imo:
          anyOf:
          - type: string
          - type: 'null'
          title: Imo
        type:
          anyOf:
          - type: string
          - type: 'null'
          title: Type
      type: object
      required:
      - id
      - organization_id
      - name
      title: VesselRead
    VesselUpdate:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 255
            minLength: 1
          - type: 'null'
          title: Name
        imo:
          anyOf:
          - type: string
            maxLength: 32
          - type: 'null'
          title: Imo
        type:
          anyOf:
          - type: string
            maxLength: 128
          - type: 'null'
          title: Type
      type: object
      title: VesselUpdate